refactor(itinerary): better organize components and slot config
